fix(release): do not restart the daemon when skipLockFileUpdate is set
