+++ packages/js/src/generators/release-version/utils/update-lock-file.ts
@@ -5,11 +5,12 @@ import {
   output,
 } from '@nx/devkit';
 import { execSync } from 'child_process';
+import { daemonClient } from 'nx/src/daemon/client/client';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
 import { getLockFileName } from 'nx/src/plugins/js/lock-file/lock-file';
 import { gte } from 'semver';
 
-export function updateLockFile(
+export async function updateLockFile(
   cwd: string,
   {
     dryRun,
@@ -28,6 +29,12 @@ export function updateLockFile(
       );
     }
     return [];
+  }
+
+  const isDaemonEnabled = daemonClient.enabled();
+  if (isDaemonEnabled) {
+    // temporarily stop the daemon, as it will error if the lock file is updated
+    await daemonClient.stop();
   }
 
   const packageManager = detectPackageManager(cwd);
@@ -71,6 +78,21 @@ export function updateLockFile(
   }
 
   execLockFileUpdate(command, cwd, env);
+
+  if (isDaemonEnabled) {
+    try {
+      await daemonClient.startInBackground();
+    } catch (e) {
+      // If the daemon fails to start, we don't want to prevent the user from continuing, so we just log the error and move on
+      if (verbose) {
+        output.warn({
+          title:
+            'Unable to restart the Nx Daemon. It will be disabled until you run "nx reset"',
+          bodyLines: [e.message],
+        });
+      }
+    }
+  }
 
   return [lockFile];
 }
